+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Airbrake Ruby Changelog
 
 * Make `notify/notify_sync` accept a block, which yields an `Airbrake::Notice`
   ([#201](https://github.com/airbrake/airbrake-ruby/pull/201))
+* Started sending `context/severity`, which is set to `error`
+  ([#202](https://github.com/airbrake/airbrake-ruby/pull/202))
 
 ### [v2.1.0][v2.1.0] (April 27, 2017)